add sub-command "create"
reorganizing codes
add sub-command "create"
