+++ package.json
@@ -4,7 +4,8 @@
   "description": "a cli tool to developping easily",
   "main": "dist/index.js",
   "scripts": {
-    "start": "npx tsc & node dist/index.js",
+    "build:start": "npx tsc & node dist/index.js",
+    "start": "node dist/index.js",
     "build": "npx tsc",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
