+++ src/bootstrap.ts
@@ -1,7 +1,8 @@
 import { Command } from 'commander';
-import * as fs from "fs";
 import * as path from "path";
 import figlet  from "figlet";
+import { createDir, createFile, listDirContents } from "./options";
+import { createCommand } from "./commands";
 
 export function bootstrap() {
     const program = new Command();
@@ -15,46 +16,23 @@ export function bootstrap() {
         .option("-l, --ls  [value]", "List directory contents")
         .option("-m, --mkdir <value>", "Create a directory")
         .option("-t, --touch <value>", "Create a file")
-        .parse(process.argv);
-
-    const options = program.opts();
-
-    async function listDirContents(filepath: string) {
-        try {
-            const files = await fs.promises.readdir(filepath);
-            const detailedFilesPromises = files.map(async (file: string) => {
-                let fileDetails = await fs.promises.lstat(path.resolve(filepath, file));
-                const { size, birthtime } = fileDetails;
-                return { filename: file, "size(KB)": size, created_at: birthtime };
-            });
-            const detailedFiles = await Promise.all(detailedFilesPromises);
-            console.table(detailedFiles);
-        } catch (error) {
-            console.error("Error occurred while reading the directory!", error);
-        }
-    }
-    function createDir(filepath: string) {
-        if (!fs.existsSync(filepath)) {
-            fs.mkdirSync(filepath);
-            console.log("The directory has been created successfully");
-        }
-    }
-    function createFile(filepath: string) {
-        fs.openSync(filepath, "w");
-        console.log("An empty file has been created");
-    }
-
-    if (options.ls) {
-        const filepath = typeof options.ls === "string" ? options.ls : __dirname;
-        listDirContents(filepath);
-    }
-    if (options.mkdir) {
-        createDir(path.resolve(__dirname, options.mkdir));
-    }
-    if (options.touch) {
-        createFile(path.resolve(__dirname, options.touch));
-    }
-    if (!process.argv.slice(2).length) {
-        program.outputHelp();
-    }
+        .action((options, command) => {
+            console.log("wycli options: ", options)
+            if (options.ls) {
+                const filepath = typeof options.ls === "string" ? options.ls : __dirname;
+                listDirContents(filepath);
+            }
+            if (options.mkdir) {
+                createDir(path.resolve(__dirname, options.mkdir));
+            }
+            if (options.touch) {
+                createFile(path.resolve(__dirname, options.touch));
+            }
+            if (!process.argv.slice(2).length) {
+                program.outputHelp();
+            }
+        })
+        .addCommand(createCommand)
+        .parse(process.argv);  // process argv
+    
 }
