+++ src/commands/index.ts
@@ -1,0 +1,4 @@
+import createCommand from "./create.command";
+
+
+export { createCommand };
