+++ src/options/index.ts
@@ -1,0 +1,5 @@
+import { listDirContents } from "./ls";
+import { createDir } from "./mkdir";
+import { createFile } from "./touch";
+
+export { createDir, listDirContents, createFile };
